+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ _Date TBD_
 ### Features Added
 * Add API methods for surveillance-related acivities. This includes upload, confirm, and reject for the bulk upload of surveillance data. It also includes create, update, and delete for singular changes to surveillance. Surveillance and associated non-conformities are returned with certified product details. 
 * Add API call to support updating certified_product meaningful_use_user counts with a CSV upload in CMS Management
+* Change behavior if certified product is marked as suspended or terminated by ONC. These statuses also result in a developer status change and require ROLE_ADMIN
 
 ### Bugs Fixed
 * Changed transition of care calculations for the EHR Certification ID to more closely match the rule.